+++ commons-core/src/main/java/com/github/jitwxs/commons/core/constant/HttpConstant.java
@@ -1,0 +1,17 @@
+package com.github.jitwxs.commons.core.constant;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+/**
+ * Http 常量
+ * @author jitwxs
+ */
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class HttpConstant {
+    public static String LOCAL_HOST = "localhost";
+
+    public static class Encoding {
+        public static String UTF8 = "UTF-8";
+    }
+}