+++ commons-web/src/main/java/com/github/jitwxs/commons/web/enums/ErrorCode.java
@@ -1,0 +1,10 @@
+package com.github.jitwxs.commons.web.enums;
+
+/**
+ * 错误码接口
+ */
+public interface ErrorCode {
+    int getCode();
+
+    String getMessage();
+}