+++ CHANGELOG.md
@@ -1,8 +1,19 @@
+### v1.3 2020-05-24
+
+- optimized packing tree
+- update core jar:
+  - add utils: BitUtils, ByteUtils
+  - add constant: HttpConstant
+- update web jar:
+  - add utils: CookieUtils
+  - add enums: ErrorCode
+
+### v1.0 ~ v1.2 2020-05-17
+
+- Fix revision question
+- Fix artifactId
+
 ### v1.0 2020-05-16
 
 - Create Projects
-
-### v1.0 2020-05-17
-
-- Fix revision question